+++ backend/src/main/java/com/erp/backend/result/dto/ResultClassResponse.java
@@ -2,8 +2,8 @@ package com.erp.backend.result.dto;
 
 public record ResultClassResponse(
         String className,
-        int studentCount,
-        int resultCount,
-        int subjectCount
+        long studentCount,
+        long resultCount,
+        long subjectCount
 ) {
 }
